PYI-813: Move JourneyResponse class over to common libs folder
